Script execution and dependency updates
1. Now script execution split into three parts:
    1. Init (runs after loading variables)
    2. pre (runs after request is loaded

2. Chore Update faker, requests-pkcs12, waitress

This is best suited for laod binary files converting/transforming data before executing.
