+++ Pipfile
@@ -9,12 +9,12 @@ jstyleson = "==0.0.2"
 requests = "==2.31.0"
 textx = "==4.0.1"
 js2py = "==0.74"
-requests-pkcs12 = "==1.22"
+requests-pkcs12 = "==1.24"
 parsys-requests-unixsocket = "==0.3.1"
 requests-aws4auth = "==1.2.3"
 requests-ntlm = "==1.2.0"
 restrictedpython = "==7.0"
-faker = "==22.7.0"
+faker = "==23.1.0"
 requests-hawk = "==1.2.1"
 pyyaml = "==6.0.1"
 toml = "==0.10.2"
@@ -22,7 +22,7 @@ msal = "==1.26.0"
 
 [dev-packages]
 python-magic = "*"
-waitress = "==2.1.1"
+waitress = "==3.0.0"
 flask = "==3.0.2"
 pyperf = "==2.6.2"
 pytest-benchmark = "*"
