+++ dothttp/__version__.py
@@ -1,1 +1,1 @@
-__version__ = '0.0.43a5'
+__version__ = '0.0.43a6'
