+++ examples/example.http
@@ -1,7 +1,7 @@
 @name("fetch 100 users, skip first 50")
 GET https://req.dothttp.dev/user
 "Authorization": "Basic dXNlcm5hbWU6cGFzc3dvcmQ="
-"content-type": "hai"
+"content-type": "application/json"
 ? ("fetch", "100") #
 ? ("skip", "50")
 ? projection, name
