+++ requirements.txt
@@ -2,12 +2,12 @@ jsonschema==4.21.1
 jstyleson==0.0.2
 textx==4.0.1
 js2py==0.74
-requests_pkcs12==1.22
+requests_pkcs12==1.24
 parsys-requests-unixsocket==0.3.1
 requests-aws4auth==1.2.3
 requests_ntlm==1.2.0
 RestrictedPython==7.0
-Faker==22.6.0
+Faker==23.1.0
 requests-hawk==1.2.1
 msal==1.26.0
 PyYaml==6.0.1
