+++ test_requirements.txt
@@ -1,2 +1,2 @@
-waitress==2.1.1
+waitress==3.0.0
 flask==3.0.2
